requirements.txt
cachetools==4.0.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
DAWG-Python==0.7.2
docopt==0.6.2
ffprobe-python==1.0.3
future==0.18.2
google-api-core==1.14.3
google-auth==1.10.0
google-cloud-speech==1.3.1
googleapis-common-protos==1.6.0
grpcio==1.26.0
idna==2.8
pkg-resources==0.0.0
protobuf==3.11.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
pymorphy2==0.8
pymorphy2-dicts==2.4.393442.3710985
python-telegram-bot==12.2.0
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.13.0
sox==1.3.7
tornado==6.0.3
urllib3==1.25.7
